requirements.txt
click==7.1.2
cycler==0.10.0
decorator==4.4.2
Faker==5.0.0
fonttools==4.19.1
glob2==0.7
imagecorruptions==1.1.2
imageio==2.9.0
imgaug==0.4.0
kiwisolver==1.3.1
lmdb
matplotlib==3.3.3
networkx==2.5
numpy==1.19.4
opencv-python==4.4.0.46
pandas==1.1.4
Pillow==8.0.1
pylatexenc==2.8
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.4
PyWavelets==1.1.1
scikit-image==0.17.2
scipy==1.5.4
Shapely==1.7.1
six==1.15.0
text-unidecode==1.3
tifffile==2020.9.3
tqdm
typer==0.3.2